#1421 moved modulesvisitor to a file &
created test for basic persisting
