+++ log4j.properties
@@ -11,4 +11,4 @@ log4j.logger.org.apache.spark.repl.SparkIMain$exprTyper=INFO
 log4j.logger.org.apache.spark.repl.SparkILoop$SparkILoopInterpreter=INFO
 
 # Un-comment to turn on smv debug log
-#log4j.logger.smv=DEBUG
+log4j.logger.smv=INFO
